Extracted some functionality to other classes for compilation
